+++ src/master/embedded/mod.rs
@@ -16,5 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+pub mod counter;
 pub mod master;
-pub mod mcaptcha;